AI_Agent/requirements.txt
crewai
flask
pyyaml
Flask-Cors